Make anonymous providers serialisable
Replaced anonymous class with ProviderProvider in ProviderSetProvider. This makes it serialisable.This improves code reuse and readability by consolidating instance retrieval logic into a dedicated class.
